requirements.txt
backcall==0.2.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.1
ipykernel==6.3.0
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==7.0.2
jupyter-core==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
nest-asyncio==1.5.1
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pyzmq==22.2.1
six==1.16.0
spyder-kernels==2.1.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
